refactor: pages 하위 디렉토리 이름 간결화
