--- src/pages/CardListPage/index.stories.js
@@ -1,11 +1,0 @@
-import React from 'react';
-import { CardListPage } from '.';
-
-export default {
-  title: 'Pages/CardListPage',
-  component: CardListPage,
-};
-
-const Template = (args) => <CardListPage {...args} />;
-
-export const _CardListPage = Template.bind({});
