+++ src/pages/ListPage/index.js
@@ -2,16 +2,16 @@ import { Card, Text, Heading } from '../../components';
 import { ROUTE } from '../../constants';
 import './style.css';
 
-export const CardListPage = (props) => {
+export const ListPage = (props) => {
   const { setRoute } = props;
   const handleAddButtonClick = () => setRoute(ROUTE.ADD_FORM);
 
   return (
-    <div className="CardListPage">
+    <div className="ListPage">
       <Heading>보유카드</Heading>
-      <button className="CardListPage__AddButton" onClick={handleAddButtonClick}>
+      <button className="ListPage__AddButton" onClick={handleAddButtonClick}>
         <Card>
-          <Text className="CardListPage__AddSign">+</Text>
+          <Text className="ListPage__AddSign">+</Text>
         </Card>
       </button>
     </div>
