+++ src/pages/ListPage/index.stories.js
@@ -1,0 +1,11 @@
+import React from 'react';
+import { ListPage } from '.';
+
+export default {
+  title: 'Pages/ListPage',
+  component: ListPage,
+};
+
+const Template = (args) => <ListPage {...args} />;
+
+export const _ListPage = Template.bind({});
